+++ cypress/support/commands.ts
@@ -1,4 +1,4 @@
-import faker from "@faker-js/faker";
+import { faker } from "@faker-js/faker";
 
 declare global {
   namespace Cypress {
