+++ cypress/tsconfig.json
@@ -4,9 +4,7 @@
     "../node_modules/@testing-library/jest-dom"
   ],
   "include": [
-    "./index.ts",
     "e2e/**/*",
-    "plugins/**/*",
     "support/**/*",
     "../node_modules/cypress",
     "../node_modules/@testing-library/cypress"
@@ -16,7 +14,7 @@
     "noEmit": true,
     "types": ["node", "cypress", "@testing-library/cypress"],
     "esModuleInterop": true,
-    "jsx": "react",
+    "jsx": "react-jsx",
     "moduleResolution": "node",
     "target": "es2019",
     "strict": true,
