+++ package.json
@@ -58,10 +58,10 @@
     "binode": "^1.0.5",
     "c8": "^7.11.3",
     "cross-env": "^7.0.3",
-    "cypress": "^9.7.0",
+    "cypress": "^10.0.3",
     "eslint": "^8.17.0",
     "eslint-config-prettier": "^8.5.0",
-    "happy-dom": "^3.2.2",
+    "happy-dom": "^5.2.0",
     "msw": "^0.42.1",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.14",
